rapiexpress/requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.0.6
Pillow==9.2.0
PyMySQL==1.0.2
sqlparse==0.4.2
